+++ development/images/wikibase/configuration/extension-loaders/CirrusSearch.php
@@ -1,8 +1,0 @@
-<?php
-
-if (isset($elasticsearchHost)) {
-  // https://www.mediawiki.org/wiki/Extension:CirrusSearch
-  wfLoadExtension( 'CirrusSearch' );
-
-  // See WikibaseCirrusSearch.php for further configuration
-}