+++ development/images/wikibase/configuration/extension-loaders/Linter.php
@@ -1,5 +1,0 @@
-<?php
-
-// https://www.mediawiki.org/wiki/Extension:Linter
-## Linter Extension (required by DiscussionTools)
-wfLoadExtension( 'Linter' );