+++ development/tests/pingback/LocalSettings.php
@@ -1,4 +1,0 @@
-<?php
-
-$wgWBRepoSettings['wikibasePingback'] = true;
-$wgWBRepoSettings['pingbackHost'] = 'http://pingback-beacon';